Added currentPage change functionality
At the bottom of pages, currentPage, re-named leftPage is shown, along with appropriate back/next buttons that I may in the future replace with arrows. Buttons are conditionally rendered on each page if the screen is only big enough to show 1 page, requiring the addition of a userWindowDimensions hook, copied from QoP on stackoverflow.