Bug fix
By default, with a small screen size, pages each provide navigation, rather than a 'back' option on the left page and a 'next' option on the right. If the screen was small enough for this case, but large enough to view part of the right page, even when on page 1, page 2's option to go 'back' was visible, thus allowing for navigation to page 0, which re-directs to not found. I removed the 'back' button in this special case.